backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "devtools>=0.12.2",
    "fastapi>=0.121.3",
    "google-adk>=1.18.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
